+++ RELEASES.md
@@ -20,6 +20,7 @@ Breaking changes
 
 * `Module::set_fn`, `Module::set_raw_fn` and `Module::set_fn_XXX_mut` all take an additional parameter of `FnNamespace`.
 * `unless` is now a reserved keyword.
+* `EvalPackage` is removed in favor of `Engine::disable_symbol`.
 
 New features
 ------------